+++ parser_lexer.py
@@ -45,10 +45,11 @@ tokens = (
     "AND",
     "OR",
     "XOR",
-    "LENGTH"
+    "LENGTH",
+    "COMPARISON"
 )
 
-
+t_COMPARISON =r"(==)|(>)|(<)|(>=)|(<=)"
 t_PLUS = r"\+"
 t_MINUS = r"-"
 t_TIMES = r"\*"
@@ -59,6 +60,7 @@ t_RPAREN = r"\)"
 t_SEMICOLON = r";"
 t_ASSIGNMENT = r"="
 t_NEG = r"!"
+
 
 reserved = {
     "if": "IF",
@@ -181,7 +183,8 @@ def p_expression_binop(p):
     | expression DIVIDE expression
     | expression AND expression
     | expression OR expression
-    | expression XOR expression"""
+    | expression XOR expression
+    | expression COMPARISON expression"""
 
     p[0] = BinOp(p.lineno(2), p[1], p[2], p[3])
